Add TransactionBase and extend from it.
